Revert "Revert "feat(@lexical/devtools): Added TreeView rendering instead of …"
This reverts commit 4fb4c8bf0ae1c3cc296008a39a9e0496b95c5020.
