+++ packages/lexical-devtools-core/src/index.ts
@@ -1,0 +1,12 @@
+/** @module @lexical/devtools-core */
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+export * from './generateContent';
+export * from './TreeView';
+export * from './useLexicalCommandsLog';
