+++ packages/lexical-devtools/package.json
@@ -27,7 +27,9 @@
     "@vitejs/plugin-react": "^4.2.1",
     "typescript": "^5.3.3",
     "lexical": "0.14.2",
+    "@lexical/devtools-core": "0.14.2",
     "wxt": "^0.17.0",
-    "vite": "^5.2.2"
+    "vite": "^5.2.2",
+    "@rollup/plugin-babel": "^6.0.4"
   }
 }
