+++ packages/lexical-devtools/src/entrypoints/content/index.ts
@@ -5,21 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  *
  */
-import {allowWindowMessaging, sendMessage} from 'webext-bridge/content-script';
+import {allowWindowMessaging} from 'webext-bridge/content-script';
 
 import useExtensionStore from '../../store';
 import storeReadyPromise from '../../store-sync/content-script';
 import injectScript from './injectScript';
 
 export default defineContentScript({
-  main(ctx) {
+  main(_ctx) {
     allowWindowMessaging('lexical-extension');
 
-    sendMessage('getTabID', null, 'background')
-      .then((tabID) => {
-        return storeReadyPromise(useExtensionStore).then(() => {
-          injectScript('/injected.js');
-        });
+    storeReadyPromise(useExtensionStore)
+      .then(() => {
+        injectScript('/injected.js');
       })
       .catch(console.error);
   },
