+++ packages/lexical-devtools/src/entrypoints/devtools-panel/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
   <body>
-    <div id="root"></div>
+    <div id="root">Loading Lexical DevTools UI...</div>
     <script type="module" src="./main.tsx"></script>
   </body>
 </html>
